Moved domain back to core
